Assignment 2/Meaning Getter/Backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
requests==2.31.0
openai==1.12.0
google-generativeai==0.3.2 